bemicro_adc: Adding second channel to display on LEDs
The signal at ADC_IN2, placed on the header at the top left side of the BeMicro
has been added to the bemicro_adc project. The conversion value is displayed on
the BeMicro's 8 LEDs using the led_pwm VHDL entity.

Selection of which channel to be displayed is done using push button 1 (SW1).