uart: Hopefully fixed the bloody RX bug
The bug was related to the baud rate generator significantly losing
sync with the DTE's baud rate generator, which resulted in bits
received from the DTE being "missed" by the half-bit tick.

=====================================================================
(Byte <x>)
                  _________________
rx                                 |_____________
                               _
baud_halfbit_tick ____________| |________________

(Byte <x+1>) -- sampling error due to baud clocks shifting
                  ___________
rx                           |___________________
                               _
baud_halfbit_tick ____________| |________________
=====================================================================

This resulted in mangled bytes at the input, an error out of which one
could not recover until the transmission from the DTE is stopped, which would
allow the baud rate generator on the UART module to stop. A subsequent
start condition from the DTE would make the baud rate be in sync again.

The issue has been solved by synchronizing the baud rate generator if
a start condition is detected while the UART module is still in the RX_STOP
state on reception. A new baud tick is issued in such a case, and both the
RX and TX FSMs go to the IDLE state from their STOP states.