i2c-wbm: Fixed tx_start_p bug
This bug resulted in too many bytes being sent over UART and thus
incorrect interpretation at the PC end. It was fixed by adding another
(temporary) state to the main FSM in the module. The job of this state
(UART_TX_START) is simply to assert the tx_start_p signal.