BTM: The SIPHRA _finally_ responds on SPI!!!
Some fixes that were needed for this:

1. The SIPHRA actually takes in 32 bits of data per register, not the variable
number of bits advertised in the alpha-state datasheet we have.
2. The order of WB registers needed to be changed, so that an SPI operation is
performed after writing the data to be relayed to the SIPHRA in a Wishbone
register. Previously this was not the case, and we were always starting an
SPI operation when writing the CSR, but the data was written to DATAR after the
operation was started. The end result was that we were always writing zeroes to
the SIPHRA, so of course we could read nothing else _but_ zeroes.
3. A 5us delay was implemented from the assertion of nCS to the start of SCLK; a
similar delay exists after the last SCLK to the de-assertion of nCS. This is
similar to what the Galao does on the SPI bus, it is still to be confirmed
(tomorrow) whether this delay is needed or not.

I thought (in desperation) that the SIPHRA SYSCLK may be needed for SPI to work,
but it turns out it isn't, the SIPHRA datasheet is at least correct and concrete
about that. :)