siphra_ctrl: Removed variable number of bits for SPI transfer
Seeing as how the ASIC actually takes in 32 bits of data for every register, the
siphra_ctrl module interface has been made simpler by removing the generics for
number of address bits and max. number of bits for data registers in the ASIC in
the siphra_ctrl module. The module now simply takes in 7-bit addresses and 32-bit
data which it applies to the SPI packet sent to the ASIC.

Software should take into account the fact that only some bits are used in the
registers, the (wb_)siphra_ctrl modules are transparent to number of bits.

The function to adapt the number of bits has also been removed from
siphra_pkg.vhd.