Renamed OBC interface component
The previous name of the component (mist_i2cs_wbm_bridge) is inaccurate
considering the planned developments for the component. The new name of the
component is "mist_obc_interface".

This has been reflected in the various places the component is used and the code
has been compiled to test that the inclusion is proper.

NB: Still no testing on the functionality of the new protocol.