Adding transaction states to master FSM
Slowly moving towards testing a full transaction based on the SET_LEDS
command. Still not tested yet, though. The FSM compiles and goes into the
RECEIVE_F_ACK state, but stays there waiting for a reception to occur. This
is however impossible, since there's no slave connected to the master.