Send UART master data at right time as part of testbench
Previously, the UART master in the testbench was sending bytes in a
delayed manner due to the nature of its FSM. Now, another state has
been added to the FSM which handles supplying as TX data to the master
the next byte in the header one clock cycle after the header has been
shifted. In this manner, the proper byte in the header is shifted out
and there are no delays in the sending of the frame.