Stopping UART wrapper when transaction header received
The pulse signal to stop the mist_uart_wrapper component is ticked after
the whole transaction header is received. This allows the whole FSM to move
forward.

To remind the reader, this stop tick signal to the wrapper is needed to mimic
an I2C stop condition and return the FSM inside the mist_uart_wrapper to IDLE.