Begin revamping OBC interface
I've decided it may be better to try for a two-FSM approach, so I'll go
ahead and try this option out. In this commit, the whole FSM has been
removed from the OBC interface component.

I think the HDL revamping should include complete removal of the Wishbone
bus from the gateware, so this commit also removes the Wishbone-specific
parts from the OBC interface component and the testbench using it.

The code has not been synthesized - that's a later step.