Make sure the OBC interface gets the LED value
There was previously a bug in the code, whereby the last byte of data in
the RX_DATA_BYTES was not received, so the in the case of the SET_LEDS
command, the LED value was not properly received. This bug was fixed
(changing the max. index of the frame_byte_count before switching back
to WAIT_I2C_ADDR) and the data is now received properly.

NB: Some hacks had to be implemented to the tesbench, as this one also
had some bugs in it on transmission. The testbench FSM is starting to
be a pile of hacks, so it needs some serious rethinking. Likely, a
reimplementation based on that in the MIST OBC interface component will
be employed.