Fix small bug/inconsistency in UART component
The UART component was found to have a small inconsistency/bug in its
rx_data_o output, which was set when the rx_ready_o pin (NB: non-pulsed!)
was high. This is inconsistent, as when the RX is ready (on the cycle
when rx_ready_o is high), the rx_data_o output should present the
received data. The RX data should not be presented one clock cycle
later, otherwise external modules using the data will receive the old
data when rx_ready_o is high -- as was discovered by using the UART module
as part of the MIST OBC interface testbench.