Begin revamping of the MIST OBC interface testbench
The beautiful new testbench uses procedures inside the stimuli process
to initiate MSP transactions and send frames within a transaction. This
is easier to understand and avoids all those nasty hacks in the previous
implementation. The downside is the simulation seems to run a bit slower
than before, but nothing significantly slow for a 3 ms run time.

In this commit, the transaction runs up to the end of RX_F_ACK
transaction state for the SET_LEDS command.