Replace Wishbone bus by new OBC interface
The BTM top-level module has now been changed to contain the new OBC
interface module, which selects peripherals depending on received
commands and sends bytes to them directly. The new interface, when
completed, should provide for a faster application of data than with
Wishbone, at the expense of Wishbone's modularity. However, for CUBES,
since it's a fairly undemanding module, it should do better with having
faster data application; the Wishbone bus previously implemented also
adds to the number of LEs used.

In this commit, only the LEDs can be set using the SET_LEDS command.
This has not yet been tested to operate in hardware, still some work
is needed on the software side for this to happen.