Rename constants in MIST OBC interface
This commits redefines constants to have "c_" prepended to the constant
names and un-capitalizes them.

Background:

Finally, I think it's best if I follow the OHWR coding guidelines
(www.ohwr.org/documents/24), since with the file structure I usually
follow to write VHDL, generics, constants, state definitions can be
found at a well-defined place in the VHDL file (generics are at the top
of the file in the entity declaration, states at the very top of the
architecture declaration and constants somewhere after the state type
definition). This all makes the code easier to read and track down.