Add parameters to receive_header procedure
This change was done to the mist_obc_interface testbench so as to be
able to return the OPCODE, FID and DL fields sent by the slave for use
in processing. The OPCODE and FID can be used for checking that the
frame sent by the slave is correct, while the DL field should be used by
the testbench to asses how many bytes the slave wants to send.