Change FID logic
The FID bit in MSP heaeders is controlled by the sender. Thus, when the
OBC sends, it controls the FID bit and CUBES stores it. When the OBC
receives (request transaction), CUBES controls the FID bit and the OBC
stores it. This has been implemented in the gateware by means of another
register (obc_send_trans) to store the transaction type. The FID bit is
only set by CUBES in TX frames if obc_send_trans = '0', i.e., this is an
OBC request transaction.

For now, OBC send transactions are assumed only and there is no error
checking involved. This means that in RX states, the FID is stored
blindly without checking. To be thought upon...

This commit also contains a modification to the wave.do file, to account
for the new FID signal being named 'fid', instead of 'rx_fid' as before.