Correct FID logic in testbench
The FID logic was not entirely correct in the testbench, due to my
understanding not being entirely correct. The FID bit should toggle
between frames as controlled by the sender. Previously, this was not the
case, the FID was controlled even by the receiver (the MIST OBC
interface, in the case of the testbench and the SET_LEDS command.) Now,
only the master controls the FID for the SET_LEDS command.

The wave.do file has also been updated to reflect the addition of
another signal (frame_end_p) to be used when the monitor process is
implemented.