Fix TID logic (and slightly re-update FID logic)
The TID signal was set on reception of the FID+OPCODE byte when
obc_send_trans was high. However, the obc_send_trans signal was set by
the trans_state FSM only after the full transaction header frame is
RXed and the OPCODE was decoded. This meant the TID signal was not being
set at all, since the OPCODE was not detected yet and obc_send_trans
could not be set.

The fix is to add the TID signal setting to the transaction FSM since
the TID is anyway a transaction-wide signal, so I think it is logical
for it to be set by the transaction FSM. The TID signal is set to the
value of the FID when the TRANS_HEADER frame has been RXed. It should be
safe by this time to set it to FID, since even in the case of frames
where CUBES TXes (and thus sets the FID), on the next clock cycle after
the frame has been RXed (frame_rxed_p = '1'), the FID should already
have been set by the frame FSM to the proper value. This latter part is
yet to be implemented, so this may change in the future.