Add LEDs sequencer after reset
The added logic sequences the BeMicro LEDs on startup, after reset has
been de-asserted. Each LED is turned on for 12.5 ms starting with D1 and
ending with D8 on the BeMicro. After D8 has been turned on, the sequence
reverses, starting with D8 and ending with D1.

The LED sequencing thus takes 200ms to complete. It is only performed on
reset to signal that the gateware is ready to operate. After the
sequence completes, control of the LEDs is relinquished to the
mist_obc_interface component.