Add internal reset to BTM
I'm intending to use the BTM module directly in the simulation, it makes
it easier to not have to reimplement everything that's inside the BTM in
the testbench. Some form of reset is required for this inside the
BTM module, and such POR logic has been implemented as part of this
commit. It's anyway better to have a POR for the logic, in case we're
missing buttons on the FPGA board.

The reset period is configurable in 50 MHz clock ticks. The
default value is 100 us.