Fix bug in UART module
The detection of the half-bit was wrong in the UART module; we weren't
actually dividing baud_div_i by 2, we were just checking that the bottom
bits of baud_div (the divider counter) matched the divide-by-2 value.
This is not an actual divide-by-2 and it came out at me when I tried to
set g_baud_div to (2^N)-1: the half-bit ticks were generated both at
half-bit and at full-bit intervals, thus resulting in the shift
registers shifting in/out too many bits.

By actually dividing the value by two and checking the whole baud_div
value vs. this divide-by-two, the half-bit ticks are issued correctly
_only_ at the half of the bit.