Make small changes to reset logic
The reset for the PLLs unlocked pulse signal, otherwise a reset would
cause a glitch on the signal, due to the fact the pulse signal is
triggering the reset it is triggering. The other change is in adding the
pulse signal to the VHDL reset process' sensitivity list, to squelch
Quartus's warning.