Add SIPHRA OBC peripheral to BTM
Currently, the SIPHRA OBC peripheral component is meant to present the
data interface for several peripherals, all addressing and hiding the
SIPHRA control component.

In order to add the component, a multiplexer for the OBC data source
interface had to be implemented. I've tested out that one can still read
the HK regs properly using this multiplexer, but the SIPHRA OBC
peripheral is still untested.