Almost make a SIPHRA_REG_OP transaction work
Several changes have been made to the BTM, OBC interface and SIPHRA OBC
peripheral modules that almost get the component in working order, at
least on the SIPHRA_REG_OP side. There are still some issues TBD, the
register data is not being shifted in correctly, but this may also be
due to the testbench. To be fixed in the following commit.