Fix siphra_obc_periph writing too many bytes to OBC buffer
The module was previously writing too many bytes to the OBC buffer, due
to the way the WE pin to the buffer was being set. This has been fixed
now by checking that the 'addr' signal is less than 'num_bytes-1' for
writing to the RAM and de-asserting the WE signal when it's less than
or equal 'num_bytes-1'. All this is done in the RECEIVE_DATA state.