Add treating of MSP NULL frames in transaction headers
Previously, NULL frames were treated as any other unwanted frame - by
sending a T_ACK. However, this doesn't seem to  be the expected behavior
according to the MSP (draft) specification. So a change was implemented
to make the transaction FSM go back to IDLE when a NULL frame is
encountered while the FSM is in its TRANS_HEADER state.

Further work needs to be done on responding to unexpected frames while
in the frame FSM is in TX_*_BYTES and RX_*_BYTES states.