Implement some error checking in the MSP
Some error handling in the MSP specification has been implemented as
part of the last two commits. Namely, the cases when the OBC asks CUBES
to put bits on the bus at the wrong time and when a header frame is
received instead of an expected data frame have been implemented.

This has been done in two steps (corresponding to commits HEAD~1 and HEAD):

1. The frame FSM checks for the I2C operation the OBC expects to perform
when a new I2C frame arrives. If a wrong type of operation is received,
the CUBES slave either decodes the received (header) frame in case it
was expecting to transmit but the OBC asks it to receive, or sends a
NULL frame in case it was expecting to receive but the OBC asks it to
transmit. This has been implemented by a couple of pulse signals that
make the transaction FSM go to specific states. See the diff between
HEAD~1 and HEAD~2 for details.

2. In case a header frame is received when a data frame is expected, the
transaction FSM is put back into the TRANS_HEADER state using one of the
two pulse signals used as part of the first point above. A delayed
version of frame_rxed_p issued by the frame FSM is used by the
transaction FSM to check the new OPCODE in this case. See the comments
of the TRANS_HEADER state of the transaction FSM for the thorough
explanation.

This commit fixes the damage done by the previous WIP commit.